Refactor: 아티스트 페이지 컴포넌트 분리
