+++ src/features/artist/components/ArtistPage.tsx
@@ -1,0 +1,55 @@
+"use client";
+
+import { SpotifyAlbum, SpotifyArtist, SpotifyTrack } from "@/lib/spotify-api";
+import { ErrorState } from "./ErrorState";
+import { AlbumList } from "./sections/AlbumList";
+import { ArtistHeader } from "./sections/ArtistHeader";
+import { ArtistInfo } from "./sections/ArtistInfo";
+import { TopTracks } from "./sections/TopTracks";
+import { ArtistSkeleton } from "./skeletons/ArtistSkeleton";
+
+interface ArtistPageProps {
+  artist: SpotifyArtist;
+  topTracks: SpotifyTrack[];
+  albums: SpotifyAlbum[];
+  isLoading?: boolean;
+  error?: string | null;
+}
+
+export function ArtistPage({
+  artist,
+  topTracks,
+  albums,
+  isLoading = false,
+  error = null,
+}: ArtistPageProps) {
+  if (isLoading) {
+    return <ArtistSkeleton />;
+  }
+
+  if (error || !artist) {
+    return <ErrorState error={error} />;
+  }
+
+  return (
+    <div className="py-6">
+      <ArtistHeader artist={artist} />
+
+      {/* 컨텐츠 영역 */}
+      <div className="container px-4 mt-12">
+        <div className="grid grid-cols-1 lg:grid-cols-3 gap-6">
+          {/* 왼쪽 및 중앙 컬럼 */}
+          <div className="lg:col-span-2 space-y-6">
+            <TopTracks tracks={topTracks} />
+            <AlbumList albums={albums} />
+          </div>
+
+          {/* 오른쪽 컬럼 */}
+          <div className="space-y-6">
+            <ArtistInfo artist={artist} />
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
