+++ src/features/artist/components/ErrorState.tsx
@@ -1,0 +1,21 @@
+"use client";
+
+import Link from "next/link";
+import React from "react";
+
+interface ErrorStateProps {
+  error: string | null;
+}
+
+export const ErrorState: React.FC<ErrorStateProps> = ({ error }) => {
+  return (
+    <div className="flex flex-col items-center justify-center h-[70vh]">
+      <h1 className="text-2xl font-bold mb-4">
+        {error || "아티스트를 찾을 수 없습니다"}
+      </h1>
+      <Link href="/" className="btn btn-primary">
+        홈으로 돌아가기
+      </Link>
+    </div>
+  );
+};
